refactor: factor out migrations into own module
