+++ src/db/mod.rs
@@ -11,6 +11,7 @@ mod config;
 pub mod criteria;
 mod date;
 pub mod init;
+mod migration;
 pub mod repository;
 
 pub fn save_all(incomes: &[Income]) -> anyhow::Result<usize> {
